requirements.txt
crewai==0.28.8
crewai-tools==0.2.6
streamlit==1.32.0
python-dotenv==1.0.0
langchain-community==0.0.38
duckduckgo-search==5.3.0
pyyaml==6.0.1
